# apps/web/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "k": {
            "default": 30,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "retrieval_mode": {
            "$ref": "#/components/schemas/RetrievalMode",
            "default": "auto"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "description": "A grounded answer, or an honest refusal \u2014 never a fluent guess.\n\n`answer` is None exactly when `insufficient_evidence` is true. `sources` carries\nthe retrieved passages so the UI can render what the answer was grounded ON, and\nevery citation's `marker` indexes into it (1-based).",
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationView"
            },
            "title": "Citations",
            "type": "array"
          },
          "insufficient_evidence": {
            "title": "Insufficient Evidence",
            "type": "boolean"
          },
          "query_tokens": {
            "title": "Query Tokens",
            "type": "integer"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalView"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SearchResultView"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "insufficient_evidence",
          "citations",
          "sources",
          "attempts",
          "query_tokens",
          "retrieval"
        ],
        "title": "AskResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Enterprise Memory OS gateway",
    "title": "JUTSU API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ask": {
      "post": {
        "description": "Answer a question from retrieved evidence (non-negotiable 3), or refuse.\n\nOrdering is the cost control, exactly as on /v1/search \u2014 with one extra gate at the\nvery front: an unconfigured answer service refuses before a single token is spent\non budget, embedding or retrieval, because \"search works, answers are not set up\"\nis a fact the caller deserves for free.\n\nThe same permission as search, deliberately: composing retrieved evidence into a\ncited paragraph grants access to nothing the caller could not already read one\npassage at a time.",
        "operationId": "ask_v1_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "search"
        ]
      }
    },
...
    "/v1/kt/{kt_code}/ask": {
      "post": {
        "description": "One turn of the KT copilot: a question answered from evidence inside the package\nwindow, with the conversation so far as context, and both turns kept.\n\nThe ordering is the cost control, the same as `/v1/ask`: the free configuration gate\nfirst, then the budget on its own committed session, then the paid embedding, then\nretrieval and synthesis. A refused caller costs nothing and learns nothing.",
        "operationId": "ask_v1_kt__kt_code__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotAskPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotTurnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/bookmarks": {
      "get": {
        "operationId": "read_bookmarks_v1_kt__kt_code__bookmarks_get",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarksOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Bookmarks",
        "tags": [
          "kt"
        ]
      },
      "post": {
        "description": "Save a claim, document, message or question. A claim or document must be inside\nthe package's gates before it is saved \u2014 the refusal is the same 404 as for an id that\nnever existed.",
        "operationId": "create_bookmark_v1_kt__kt_code__bookmarks_post",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Bookmark",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/bookmarks/{bookmark_id}": {
      "delete": {
        "operationId": "delete_bookmark_v1_kt__kt_code__bookmarks__bookmark_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bookmark Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bookmark",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/conversations": {
      "get": {
        "operationId": "read_conversations_v1_kt__kt_code__conversations_get",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Conversations",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/conversations/search": {
      "post": {
        "description": "Find earlier conversations by what was said in them. POST, so the words stay\nout of the URL.",
        "operationId": "search_conversations_v1_kt__kt_code__conversations_search_post",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationSearchPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Conversations",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/conversations/{conversation_id}": {
      "get": {
        "description": "A conversation with its turns. Citations are re-checked against the package as it\nis now; a cited document no longer inside it renders unavailable.",
        "operationId": "read_one_conversation_v1_kt__kt_code__conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read One Conversation",
        "tags": [
          "kt"
        ]
      }
    },
    "/v1/kt/{kt_code}/conversations/{conversation_id}/archive": {
      "post": {
        "operationId": "archive_one_conversation_v1_kt__kt_code__conversations__conversation_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive One Conversation",
        "tags": [
          "kt"
        ]
      }
    },
...
    "/v1/kt/{kt_code}/progress": {
      "get": {
        "operationId": "read_progress_v1_kt__kt_code__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "kt_code",
            "required": true,
            "schema": {
              "title": "Kt Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Progress",
        "tags": [
          "kt"
        ]
      }
    },
...
  }
}